+++ src/panels/header/AppHeader.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: computerPixel;
-  src: url("../assets/fonts/ComputerPixel/computer_pixel-7.ttf");
+  src: url("../../assets/fonts/computer_pixel-7.ttf");
 }
 
 .title span {
