+++ src/panels/resources-panel/ResourcePanel.css
