+++ src/shared/GameState.ts
